+++ src/Timestampable/UpdatedAt.php
@@ -18,7 +18,7 @@ trait UpdatedAt
      *
      * @ORM\Column(type="datetime", name="updated_at", nullable=true)
      */
-    #[Column(type: "datetime", name: "updated_at", nullable: true)]
+    #[ORM\Column(type: "datetime", name: "updated_at", nullable: true)]
     protected ?\DateTime $updatedAt;