+++ src/TimestampableImmutable/ConnectedAt.php
@@ -18,7 +18,7 @@ trait ConnectedAt
      *
      * @ORM\Column(type="datetime", name="connected_at", nullable=true)
      */
-    #[Column(type: "datetime", name: "connected_at", nullable: true)]
+    #[ORM\Column(type: "datetime", name: "connected_at", nullable: true)]
     protected ?\DateTimeImmutable $connectedAt;